fix author-select and libraries tests
